gen-dashboard.py: limit fstests results
We use kdevops and kdevops-kpd to test fstests, but we don't want
those tests to be included as real filesystem tests. These kdevops
and kdevops-kpd fstests are designed to just test to ensure kdevops
will always work. They just run generic/003 so to make sure that
we can:

  * build linux
  * build fstests
  * run generic/003

That's it. So ignore those tests from the official filesystem pane.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>